% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.r
\name{fn_rpart}
\alias{fn_rpart}
\title{Simple package to build a RPART classifier for the Bank sales data.}
\usage{
fn_rpart()
}
\value{
Saved model file
}
\description{
Simple package to build a RPART classifier for the Bank sales data.
}
\examples{
fn_rpart()
}
